refactor: shared/ui/atom 디렉토리 구조 통일성있게 수정 (#78)
