+++ src/shared/ui/atoms/button/variant/IconButton.tsx
@@ -1,6 +1,6 @@
 import clsx from "clsx";
 import { Icon } from "@shared/ui/atoms";
-import type { IconName } from "@shared/ui/atoms/icon/variants";
+import type { IconName } from "@shared/ui/atoms/icon/variant";
 import { BaseButton, type BaseButtonProps, type ButtonContainerPalette } from "../foundation";
 
 interface Palette extends ButtonContainerPalette {
