+++ src/shared/ui/atoms/icon/variant/base.ts
