+++ src/shared/ui/atoms/icon/variant/brand.ts
