+++ src/shared/ui/atoms/icon/variant/index.ts
