+++ src/shared/ui/atoms/input/variant/UnderlineInput.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import { Pressable, TextInput, type TextInputProps, View } from "react-native";
 import { Icon } from "@shared/ui/atoms";
-import type { IconName } from "@shared/ui/atoms/icon/variants";
+import type { IconName } from "@shared/ui/atoms/icon/variant";
 import { type InputRequiredProps, InputStyle } from "../foundation";
 import ClearButton from "../part/ClearButton";
